@@ -4,12 +4,11 @@ homepage: https://github.com/tiagohm/cuber
 version: 0.2.0
 
 environment:
-  sdk: ">=2.8.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
-  meta: ^1.2.0
-  equatable: ^1.2.0
+  equatable: ^2.0.3
 
 dev_dependencies:
-  test: ^1.15.0
-  pedantic: ^1.9.0
+  test: ^1.17.9
+  pedantic: ^1.11.1
